Add CITATION.cff
This adds a CFF file to the superdataset. The file can be used as a
source of metadata for ZENODO (e.g. listing authors). It can also
provide DataLad extractors with information about DOI (not available
from studyminimeta).

Compared to the studyminimeta file, it has the DOI but lacks funding
property. Other than that, the two files mostly duplicateeach other's
information. It's fine to have them coexist, although CFF could also
be left as the only one, since it is a widely adopted standard (while
studyminimeta was created mostly for local archival).